# added an example to call performance command.