--- a/magellan.json
+++ b/magellan.json
@@ -1,0 +1,5 @@
+{
+  "reporters": [
+    "./node_modules/testarmada-performance-reporter/lib/performance_reporter"
+  ]
+}