--- a/package.json
+++ b/package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "nightwatch": "^0.7.9",
     "testarmada-magellan": "1.6.6",
-    "testarmada-magellan-nightwatch": "1.5.1"
+    "testarmada-magellan-nightwatch": "1.5.1",
+    "testarmada-performance-reporter": "1.1.1"
   }
 }