--- a/tests/walmart-book.js
+++ b/tests/walmart-book.js
@@ -1,11 +1,19 @@
 var Test = require("../lib/example-base-test-class");
-
+var url = "http://www.walmart.com/search/?query=sam%20walton%20made%20in%20america";
 module.exports = new Test({
 
   "Search for Sam Walton Book": function (client) {
     client
       .resizeWindow(1280, 1024)
-      .url("http://www.walmart.com/search/?query=sam%20walton%20made%20in%20america")
+      .url(url)
+      .getPerformance(function (result) {
+        // Send a message to Magellan about this page
+        process.send({
+          type: "performance-metrics",
+          url: url,
+          metrics: result
+        });
+      })
   },
 
   "Check product description": function (client) {