This can be executed with command `magellan --browser firefox` . Chrome is sometimes giving me negative numbers, still working on it.  This is example requetsed  for PR #18 in magellan-nightwatch repo. Here is how the result should look like 
```
============= Suite Complete =============

     Status: PASSED
    Runtime: 14.4s
Total tests: 2
 Successful: 2 / 2

---------------------------------------
Gathered performance metrics: 
---------------------------------------

url: http://www.walmart.com/search/?query=sam%20walton%20made%20in%20america
metrics:
[
  "DomContentLoad: 0.525 seconds",
  "Page Load (onLoad): 3.973 seconds",
  "Full Page Load: 3.16 seconds",
  "Number of Requests: 92",
  "First Paint: 0 seconds"

]
```