@Maciek416 2015-09-22
Overriding CI results because Safari7 results are masking what is otherwise a green result. I am merging this.